fix bookings Form for listing owner

+++ frontend/src/components/renderShowListing.ts
@@ -1,6 +1,7 @@
 import { navigate } from "../main";
 import { IListing } from "../interfaces/listing";
 import axios from "axios";
+import { confirmDialog } from "../utils/confirmDialogBox";
 
 export function renderShowListing(container: HTMLElement, listing: IListing) {
   const images = listing.images;
@@ -78,17 +79,21 @@ export function renderShowListing(container: HTMLElement, listing: IListing) {
   if (deleteForm) {
     deleteForm.addEventListener("submit", async (event) => {
       event.preventDefault();
-      console.log(listing.id);
-      try {
-        await axios.delete(`/api/listings/${listing.id}`, {
-          headers: {
-            Authorization: `Bearer ${token}`,
-          },
-        });
-        navigate(`/listings`);
-      } catch (error) {
-        console.error("Error deleting listing:", error);
-        alert("Failed to delete listing.");
+      const confirmed = await confirmDialog(
+        "Are you sure you want to delete this listing?"
+      );
+      if (confirmed) {
+        try {
+          await axios.delete(`/api/listings/${listing.id}`, {
+            headers: {
+              Authorization: `Bearer ${token}`,
+            },
+          });
+          navigate(`/listings`);
+        } catch (error) {
+          console.error("Error deleting listing:", error);
+          alert("Failed to delete listing.");
+        }
       }
     });
   }
